% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_coverage.R
\name{check_coverage}
\alias{check_coverage}
\title{Check test coverage for a function}
\usage{
check_coverage(function_name, test_file = NULL)
}
\arguments{
\item{function_name}{A string of the function name (with or without \verb{()})
whose script is in the \verb{R/} folder.}

\item{test_file}{A string specifying the test script to use.
Defaults to \code{NULL}, which makes the function search for a script in the
\code{tests/testthat} folder.
Additionally, the script is assumed to have the same name as \code{function_name}
but start with \verb{test-}, and all underscores are substituted for dashes.}
}
\value{
A coverage report in the Viewer pane.
}
\description{
Check test coverage for a function
}
