% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanup-datetime.R
\name{cleanup_date}
\alias{cleanup_date}
\title{Cleans up date from data frame}
\usage{
cleanup_date(data, prefix = "", suffix = "", expand = c("Year", "Month",
  "Day"))
}
\arguments{
\item{data}{data.frame with columns of [prefix]expand[suffix], [prefix]expand[suffix] values etc}

\item{prefix}{string}

\item{suffix}{string}

\item{expand}{character vector of column names where [prefix]expand[suffix]
and expand must only be Year, Month, Day}
}
\value{
Cleaned up data
}
\description{
Cleans up date from data frame
}
\examples{
data <- data.frame(Year = 2000, Month = 1:4, Day = 2, Comments = 'No way')
cleanup_date(data)
}
\seealso{
\code{\link{cleanup_datetime}}
}

