% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanup-datetime.R
\name{cleanup_datetime}
\alias{cleanup_datetime}
\title{Cleans up datetime from data frame}
\usage{
cleanup_datetime(data, prefix = "^", suffix = "$", expand = c("Year",
  "Month", "Day", "Hour", "Minute", "Second"))
}
\arguments{
\item{data}{data.frame with columns of [prefix]expand[suffix], [prefix]expand[suffix] values etc}

\item{prefix}{string for matching regexp pattern start}

\item{suffix}{string for matching regexp pattern end}

\item{expand}{character vector of column names where [prefix]expand[suffix]
and expand must only be Year, Month, Day, Hour, Minute and/or Second
\code{\link{grep}} function}
}
\value{
Cleaned up data
}
\description{
Cleans up datetime from data frame
}
\examples{
data <- data.frame(Year = 2000, Month = 1:4, Day = 2)
cleanup_datetime(data)
data <- data.frame(Year = 2000, Month = 1:4, Day = 2, Comments = 'No way')
cleanup_datetime(data)
data <- data.frame(YearRelease = 2000, MonthReleased = 1:4, DayDeceased = 2,
                   XDay = 3, TDayY = 4)
cleanup_datetime(data)
cleanup_datetime(data, '', '')
cleanup_datetime(data, '^X')
cleanup_datetime(data, '^T')
cleanup_datetime(data, '^T', '')
cleanup_datetime(data, '^T', 'Y$')
cleanup_datetime(data, ,'Release[d]*|Deceased$')

}
\seealso{
\code{\link{extract_datetime}}
}

