% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanup-datetime.R
\name{cleanup_time}
\alias{cleanup_time}
\title{Cleans up time from data frame}
\usage{
cleanup_time(data, prefix = "", suffix = "", expand = c("Hour", "Minute",
  "Second"))
}
\arguments{
\item{data}{data.frame with columns of [prefix]expand[suffix], [prefix]expand[suffix] values etc}

\item{prefix}{string}

\item{suffix}{string}

\item{expand}{character vector of column names where [prefix]expand[suffix]
and expand must only be Hour, Minute and/or Second}
}
\value{
Cleaned up data
}
\description{
Cleans up time from data frame
}
\seealso{
\code{\link{cleanup_datetime}}
}

