% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage_summary.R
\name{coverage_summary}
\alias{coverage_summary}
\title{Check the test coverage for multiple functions}
\usage{
coverage_summary(function_files, test_files = NULL, ..., return_all = FALSE)
}
\arguments{
\item{function_files}{A character vector of file names, which are assumed to
be in the \verb{R/} folder.}

\item{test_files}{A character vector of length equal to that of
\code{function_files} specifying the test script(s) to use.
Defaults to \code{NULL}, which makes the function search for a script in the
\code{tests/testthat} folder.
Additionally, the script is assumed to have the same name as \code{function_files}
but start with \verb{test-}, and all underscores are substituted for dashes.}

\item{return_all}{A flag indicating whether the function should return all
files (including those with 100\% coverage) or only the files without complete
coverage (default).}
}
\value{
A coverage report in the Viewer pane.
}
\description{
Check the test coverage for multiple functions
}
