% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cput.R
\name{cput}
\alias{cput}
\title{Write an Object to the Clipboard}
\usage{
cput(x = NULL, envir = parent.frame())
}
\arguments{
\item{x}{An object or NULL if R code in the clipboard is to be evaluated and then write}

\item{envir}{An environment in which the contents of the clipboard are to be evaluated if x is \code{NULL}.}
}
\value{
An invisible copy to the object.
}
\description{
Writes an object (or the evaluated contents of the clipboard) to the clipboard.
}
\details{
Useful for pasting relatively simple objects into examples or tests.
}
\examples{
cput(data.frame(x = 1))
}
\seealso{
\code{\link[=dput]{dput()}}
}
