% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-datetime.R
\name{extract_time}
\alias{extract_time}
\title{Extracts time from data frame}
\usage{
extract_time(data, prefix = "", suffix = "", expand = c("Hour", "Minute",
  "Second"), tz = "PST", standardised_offset = -8)
}
\arguments{
\item{data}{data.frame with columns of [prefix]expand[suffix], [prefix]expand[suffix] values etc}

\item{prefix}{string}

\item{suffix}{string}

\item{expand}{character vector of column names where [prefix]expand[suffix]
and expand must only be Hour, Minute and/or Second}

\item{tz}{string of time zone}

\item{standardised_offset}{an integer scalar indicating the time difference
in hours between UTC and the desired time zone.}
}
\value{
A datetime object
}
\description{
Extracts time from data frame
}

