% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fold-change.R
\name{fold2nfold}
\alias{fold2nfold}
\title{Fold to n-Fold Change}
\usage{
fold2nfold(x)
}
\arguments{
\item{x}{A vector of the n-fold values}
}
\value{
A vector of the fold change
}
\description{
Fold to n-Fold Change
}
\examples{
fold2nfold(3)
}
