% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fold-change.R
\name{fold2prop}
\alias{fold2prop}
\title{Fold Change to Proportional Change}
\usage{
fold2prop(x)
}
\arguments{
\item{x}{A numeric vector of the proportional change}
}
\value{
A numeric vector of the n-fold change
}
\description{
Fold Change to Proportional Change
}
\examples{
fold_change(3, c(3, 1, 9))
prop_change(3, c(3, 1, 9))
fold2prop(fold_change(3, c(3, 1, 9)))
}
