% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fold-change.R
\name{fold_change}
\alias{fold_change}
\title{Fold Change}
\usage{
fold_change(x, y)
}
\arguments{
\item{x}{A vector of the before values}

\item{y}{A vector of the after values}
}
\value{
A vector of the fold change
}
\description{
Fold Change
}
\examples{
fold_change(3, c(3, 1, 9))
}
