% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hex-to-dec-code.R
\name{hex_to_dec_code}
\alias{hex_to_dec_code}
\title{Hexadecimal to Decimal Code Converter}
\usage{
hex_to_dec_code(hex)
}
\arguments{
\item{hex}{A string of the hex code.}
}
\value{
A double.
}
\description{
Hexadecimal to Decimal Code Converter
}
\details{
The function will return NA if non HEX letters are present. The
only letters that can be present in the HEX codes are A to F.
}
\examples{
hex_to_dec_code("349EA72A50")
hex_to_dec_code("14A5D0BE89")
hex_to_dec_code(c("14D", "E67"))
}
