% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/named.R
\name{is.named}
\alias{is.named}
\title{Test if named}
\usage{
is.named(x)
}
\arguments{
\item{x}{the object to test}
}
\value{
A logical scalar indicating whether named
}
\description{
Test if named
}
\examples{
is.named(1)
}
