% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fold-change.R
\name{nfold2fold}
\alias{nfold2fold}
\title{Fold to n-Fold Change}
\usage{
nfold2fold(x)
}
\arguments{
\item{x}{A vector of the n-fold values}
}
\value{
A vector of the fold change
}
\description{
Fold to n-Fold Change
}
\examples{
nfold2fold(3)
}
