% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fold-change.R
\name{nfold2prop}
\alias{nfold2prop}
\title{n-Fold Change to Proportional Change}
\usage{
nfold2prop(x)
}
\arguments{
\item{x}{A numeric vector of the proportional change}
}
\value{
A numeric vector of the n-fold change
}
\description{
n-Fold Change to Proportional Change
}
\examples{
nfold_change(3, c(3, 1, 9))
prop_change(3, c(3, 1, 9))
nfold2prop(nfold_change(3, c(3, 1, 9)))
}
