% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poisutils-package.R
\docType{package}
\name{poisutils-package}
\alias{poisutils}
\alias{poisutils-package}
\title{poisutils: Poisson Consulting's Utility Functions}
\description{
Utility functions for Poisson Consulting's scripts and packages.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/poissonconsulting/poisutils}
  \item Report bugs at \url{https://github.com/poissonconsulting/poisutils/issues}
}

}
\author{
\strong{Maintainer}: Joe Thorley \email{joe@poissonconsulting.ca}

Other contributors:
\itemize{
  \item Ayla Pearson \email{ayla@poissonconsulting.ca} (\href{https://orcid.org/0000-0001-7388-1222}{ORCID}) [contributor]
}

}
\keyword{internal}
