% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{poisutils}
\alias{poisutils}
\alias{package-poisutils}
\alias{poisutils-package}
\title{poisutils}
\description{
Utility functions for Poisson Consulting's scripts and packages.
}
\seealso{
\url{https://github.com/poissonconsulting/poisutils}
}
