% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fold-change.R
\name{prop2nfold}
\alias{prop2nfold}
\title{Proportional Change to n-Fold Change}
\usage{
prop2nfold(x)
}
\arguments{
\item{x}{A numeric vector of the proportional change}
}
\value{
A numeric vector of the n-fold change
}
\description{
Proportional Change to n-Fold Change
}
\examples{
nfold_change(3, c(3, 1, 9))
prop_change(3, c(3, 1, 9))
prop2nfold(prop_change(3, c(3, 1, 9)))
}
