% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fold-change.R
\name{prop_change}
\alias{prop_change}
\title{Proportional Change}
\usage{
prop_change(x, y)
}
\arguments{
\item{x}{A vector of the before values}

\item{y}{A vector of the after values}
}
\value{
A vector of the proportional change
}
\description{
Proportional Change
}
\examples{
prop_change(3, c(3, 1, 9))
}
