% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-dir.R
\name{ps_create_dir}
\alias{ps_create_dir}
\title{Create Directory}
\usage{
ps_create_dir(dir, ask = TRUE)
}
\arguments{
\item{dir}{A string of the directory name}

\item{ask}{A flag indicating whether to ask before creating a new directory.}
}
\value{
An invisible flag indicating whether the directory exists.
}
\description{
Create Directory
}
