% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/messages.R
\name{ps_error}
\alias{ps_error}
\title{Error Message}
\usage{
ps_error(..., call. = FALSE)
}
\arguments{
\item{...}{zero or more objects which can be coerced to character
    (and which are pasted together with no separator) or a single
    condition object.}

\item{call.}{logical, indicating if the call should become part of the
    error message.}
}
\description{
Stops execution with an error message.
}
