% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h2d-tag.R, R/hex-to-dec-pit-tag.R
\name{ps_h2d_tag}
\alias{ps_h2d_tag}
\title{Convert Hexidecimal to Decimal PIT Tags}
\usage{
ps_h2d_tag(hex, country_code = 900)

ps_h2d_tag(hex, country_code = 900)
}
\arguments{
\item{hex}{A string of the hex code.}

\item{country_code}{Number of the country code to add to the decimal code.}
}
\value{
A string.

A string.
}
\description{
Convert Hexidecimal to Decimal PIT Tags

Convert Hexidecimal to Decimal PIT Tags
}
\details{
If the decimal code is 11 digits long a zero is padded to the front
of the value before the country code is added. If the code is 10 or less
digits long then NA will be returned. Any decimal codes 13 or more digits
long will return NA. If the code is 12 digits long the country code is
appended directly to the decimal value.

If the decimal code is 11 digits long a zero is padded to the front
of the value before the country code is added. If the code is 10 digits or
less then \code{NA} will be returned. Any decimal codes 13 or more digits long
will return \code{NA}. If the code is 12 digits long the country code is
appended directly to the decimal value.

All values output should be 15 digits long.
}
\examples{
ps_h2d_tag("349EA72A50")
ps_h2d_tag("349EA70")
ps_h2d_tag("349EA72A50000")
ps_h2d_tag("349EA72A50", country_code = 124)
ps_h2d_tag("349EA72A50")
ps_h2d_tag("349EA70")
ps_h2d_tag("349EA72A50000")
ps_h2d_tag("349EA72A50", country_code = 124)
ps_h2d_tag(c("349EA72A50", "349EA72A49"))
}
