% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hex-to-dec-pit-tag.R
\name{ps_hex_to_dec_pit_tag}
\alias{ps_hex_to_dec_pit_tag}
\title{Convert Hexidecimal to Decimal PIT Tags}
\usage{
ps_hex_to_dec_pit_tag(hex, country_code = 900)
}
\arguments{
\item{hex}{A string of the hex code.}

\item{country_code}{Number of the country code to add to the decimal code.}
}
\value{
A string.
}
\description{
Convert Hexidecimal to Decimal PIT Tags
}
\examples{
ps_hex_to_dec_pit_tag("349EA72A50")
ps_hex_to_dec_pit_tag("349EA70")
ps_hex_to_dec_pit_tag("349EA72A50000")
ps_hex_to_dec_pit_tag("349EA72A50", country_code = 124)
}
