% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/messages.R
\name{ps_message}
\alias{ps_message}
\title{Diagnostic Message}
\usage{
ps_message(...)
}
\arguments{
\item{...}{zero or more objects which can be coerced to character
    (and which are pasted together with no separator) or (for
    \code{message} only) a single condition object.}
}
\description{
Generates a diagnostic message.
}
