% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change.R
\name{ps_nfold_change}
\alias{ps_nfold_change}
\title{n-Fold Change}
\usage{
ps_nfold_change(x, y)
}
\arguments{
\item{x}{A vector of the before values}

\item{y}{A vector of the after values}
}
\value{
A vector of the n-fold change
}
\description{
n-Fold Change
}
\examples{
ps_nfold_change(3, c(3, 1, 9))
}
