% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/punctuate.R
\name{ps_plural}
\alias{ps_plural}
\title{Plural}
\usage{
ps_plural(x, n = 1L, end = "")
}
\arguments{
\item{x}{A string to pluralize.}

\item{n}{A count of the number of occurrences of x.}

\item{end}{A string of the last part of the pluralization
(useful for adding punctuation).}
}
\value{
A string
}
\description{
Plural
}
\examples{
ps_plural("column", 1)
ps_plural("column", 2)
ps_plural("column", 3, end = ".")
}
