% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change.R
\name{ps_prop2nfold_change}
\alias{ps_prop2nfold_change}
\title{Proportional Change to n-Fold Change}
\usage{
ps_prop2nfold_change(x)
}
\arguments{
\item{x}{A numeric vector of the proportional change}
}
\value{
A numeric vector of the n-fold change
}
\description{
Proportional Change to n-Fold Change
}
\examples{
ps_nfold_change(3, c(3, 1, 9))
ps_prop_change(3, c(3, 1, 9))
ps_prop2nfold_change(ps_prop_change(3, c(3, 1, 9)))
}
