% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/punctuate.R
\name{ps_punctuate}
\alias{ps_punctuate}
\title{Punctuate}
\usage{
ps_punctuate(x, qualifier = "or")
}
\arguments{
\item{x}{A vector to punctuate}

\item{qualifier}{A string to separate the last element.}
}
\value{
A string.
}
\description{
Punctuate
}
\examples{
ps_punctuate(1:3)
}
