% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename-object.R
\name{ps_rename_object}
\alias{ps_rename_object}
\title{Rename Object}
\usage{
ps_rename_object(x, new_name, envir = parent.frame())
}
\arguments{
\item{x}{The object to rename.}

\item{new_name}{A string of the new name.}

\item{envir}{The environment.}
}
\description{
Rename Object
}
