% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round-preserve-sum.R
\name{ps_round_preserve_sum}
\alias{ps_round_preserve_sum}
\title{Round Values while Preserving their Rounded Sum}
\source{
\url{http://biostatmatt.com/archives/2902}
}
\usage{
ps_round_preserve_sum(x, digits = 0L)
}
\arguments{
\item{x}{A numeric vector of values to round}

\item{digits}{A count of the number of decimals for rounding}
}
\description{
Round Values while Preserving their Rounded Sum
}
\examples{
sum(c(0.33, 0.33, 0.33))
round(c(0.33, 0.33, 0.33), 1)
sum(round(c(0.33, 0.33, 0.33), 1))
ps_round_preserve_sum(c(0.33, 0.33, 0.33), 1)
sum(ps_round_preserve_sum(c(0.33, 0.33, 0.33), 1))
}
