% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sma.R
\name{ps_sma}
\alias{ps_sma}
\title{Simple Moving Average}
\usage{
ps_sma(x, n = 1L, na.rm = FALSE)
}
\arguments{
\item{x}{An integer or double vector.}

\item{n}{A positive integer of the window before and after each value.}

\item{na.rm}{A flag indicating whether to ignore missing values when calculating the mean.}
}
\value{
A double vector.
}
\description{
Returns the simple moving average with a window of \code{2 * n}.
}
\examples{
ps_sma(1:4)
}
