% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/messages.R
\name{ps_warning}
\alias{ps_warning}
\title{Warning Message}
\usage{
ps_warning(..., call. = FALSE)
}
\arguments{
\item{...}{\emph{either} zero or more objects which can be coerced
    to character (and which are pasted together with no separator)
    \emph{or} a single \link[base]{condition} object.}

\item{call.}{logical, indicating if the call should become part of the
    warning message.}
}
\description{
Generates a warning message.
}
