% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utc-offset.R
\name{utc_offset}
\alias{utc_offset}
\title{UTC Offset}
\usage{
utc_offset(x)
}
\arguments{
\item{x}{a date-time object}
}
\value{
The offsets as an numeric vector.
}
\description{
Calculates the offsets in hours between the time zones
of a date time object and the Coordinated Universal Time time zone.
}
\seealso{
\code{\link{standardise_datetime}}
}

