% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cput.R
\name{cput}
\alias{cput}
\title{Write an Object to the Clipboard}
\usage{
cput(x, allow_non_interactive = FALSE)
}
\arguments{
\item{x}{An object.}

\item{allow_non_interactive}{A flag specifying whether to allow
writing to the clipboard in an non-interactive session.}
}
\value{
An invisible copy to the object.
}
\description{
Write an Object to the Clipboard
}
\seealso{
\code{\link[=dput]{dput()}}
}
