% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h2d-tag.R
\name{ps_h2d_tag}
\alias{ps_h2d_tag}
\title{Convert Hexidecimal to Decimal PIT Tags}
\usage{
ps_h2d_tag(hex, country_code = 900)
}
\arguments{
\item{hex}{A string of the hex code.}

\item{country_code}{Number of the country code to add to the decimal code.}
}
\value{
A string.
}
\description{
Convert Hexidecimal to Decimal PIT Tags
}
\details{
Any decimal codes 13 or more digits long will return NA.  If the
code is 12 digits long the country code is appended directly to the decimal
value. If the decimal code is 11 digits or less then the decimal code is
padded with zeros on the front to ensure the code is 12 digits long before
appending the country code to the code.
}
\examples{
ps_h2d_tag("349EA72A50")
ps_h2d_tag("349EA70")
ps_h2d_tag("349EA72A50000")
ps_h2d_tag("349EA72A50", country_code = 124)
}
