% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-hobo-csv.R
\name{read_hobo_csv}
\alias{read_hobo_csv}
\title{Read Hobo CSVs}
\usage{
read_hobo_csv(file = ".", recursive = FALSE)
}
\arguments{
\item{file}{A string of the file or directory.}

\item{recursive}{A flag indicating whether to read files from subdirectories.
Ignore if dir is a file.}
}
\value{
A tibble of the data.
}
\description{
Read Hobo CSVs
}
\examples{
read_hobo_csv(system.file("hobo", "10723440.csv", package = "poisutils"))
}

